+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Changelog
 
 
+### v0.4.0
+
+- fullnextsearch -> fulltextsearch
+- Pagination
+- settings panel
+
+
+
 ### v0.3.2
 
 - UI: remove personal settings
 - DB: fill err field on new indexes
+
+
 
 ### v0.3.1